Add functional game channel in ActionCable
