+++ test/channels/game_channel_test.rb
@@ -1,6 +1,6 @@
 require "test_helper"
 
-class RoomsChannelTest < ActionCable::Channel::TestCase
+class GameChannelTest < ActionCable::Channel::TestCase
   # test "subscribes" do
   #   subscribe
   #   assert subscription.confirmed?
